requirements.txt
# requirements.txt
langchain
langchain-openai
faiss-cpu
pydantic
tenacity
langchain-community
pyyaml